Add files via upload
This project was organized to receive weather satellite signals from
the NOAA (National Oceanic and Atmospheric Administration) using a
Quadrifilar Helix (QFH) antenna and decode the received signals into
weather images. The involved designing and constructing a QFH
antenna and a ground station setup for receiving weather satellite
signals from the NOAA.The project also involved learning about the
technical details of the weather satellite signals and the decoding
process.